Some client networking deduplication
