+++ common/src/main/kotlin/net/stellarica/common/networking/ObjectListener.kt
@@ -1,0 +1,4 @@
+package net.stellarica.common.networking
+
+interface ObjectListener {
+}
