+++ common/src/main/kotlin/net/stellarica/common/networking/PacketListener.kt
@@ -1,0 +1,11 @@
+package net.stellarica.common.networking
+
+interface PacketListener {
+	val handler: NetworkHandler<PacketListener>
+	fun unregister() {
+		handler.unregister(this)
+	}
+	fun register() {
+		handler.register(this)
+	}
+}
