+++ server/src/main/kotlin/net/stellarica/server/craft/starship/subsystem/Subsystem.kt
@@ -1,0 +1,4 @@
+package net.stellarica.server.craft.starship.subsystem
+
+interface Subsystem {
+}
